rpk cluster loggers: new subcommand group
Move log level management from `rpk redpanda
admin config log-level set` into a new `rpk
cluster loggers` group with `list` and `set`
subcommands. The old `--help-loggers` flag
becomes the standalone `list` command.

Replace the confusing `--host` flag (which
accepted either a hostname or an admin index)
with `--node-id`. On `set`, --node-id is now
required via cobra; on `list` it is optional
and routes to any broker when unset.

The legacy `rpk redpanda admin config
log-level` parent and its `set` child are
hidden and deprecated; the inner set body is
now `loggers.NewSetCommand(fs, p)` so the
old helpers and tests are gone.
